requirements.txt
instagrapi==2.1.2
pandas==2.2.3
Pillow==8.1.1
moviepy==1.0.3
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
datetime

